# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api.K8sSecretInfo:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      created_at:
        type: string
      data_keys:
        items:
          type: string
        type: array
      labels:
        additionalProperties:
          type: string
        type: object
      managed_by_gitops:
        type: boolean
      name:
        type: string
      namespace:
        type: string
      type:
        type: string
    type: object
  internal_api.K8sSecretsListResponse:
    properties:
      namespace:
        type: string
      secrets:
        items:
          $ref: '#/definitions/internal_api.K8sSecretInfo'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/yourorg/secret-manager/issues
  description: Production-ready secret management with drift detection and GitOps
    integration
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/yourorg/secret-manager
  title: Secret Manager API
  version: "1.0"
paths:
...
  /drift-events/{drift_id}/import-to-git:
    post:
      consumes:
      - application/json
      description: Resolves drift by committing Kubernetes secret version to Git
      parameters:
      - description: Drift Event ID (UUID)
        in: path
        name: drift_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Import successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid drift event ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Drift detector not available
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import from K8s to Git
      tags:
      - drift
  /drift-events/{drift_id}/mark-resolved:
    post:
      consumes:
      - application/json
      description: Mark drift event as resolved without syncing (manual resolution)
      parameters:
      - description: Drift Event ID (UUID)
        in: path
        name: drift_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Drift marked as resolved
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid drift event ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Drift detector not available
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark drift as resolved
      tags:
      - drift
  /drift-events/{drift_id}/sync-from-git:
    post:
      consumes:
      - application/json
      description: Resolves drift by overwriting Kubernetes secret with Git version
      parameters:
      - description: Drift Event ID (UUID)
        in: path
        name: drift_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sync successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid drift event ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Drift detector not available
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Sync from Git to K8s
      tags:
      - drift
...
  /namespaces/{namespace}/k8s-secrets:
    get:
      consumes:
      - application/json
      description: List all secrets currently deployed in the Kubernetes namespace
        (metadata only, no values)
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Kubernetes secrets
          schema:
            $ref: '#/definitions/internal_api.K8sSecretsListResponse'
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Kubernetes cluster not available
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Kubernetes secrets
      tags:
      - kubernetes
  /namespaces/{namespace}/k8s-secrets/{name}:
    get:
      consumes:
      - application/json
      description: Get metadata for a specific Kubernetes secret (keys only, no values)
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      - description: Secret name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Kubernetes secret metadata
          schema:
            $ref: '#/definitions/internal_api.K8sSecretInfo'
        "404":
          description: Namespace or secret not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Kubernetes cluster not available
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Kubernetes secret details
      tags:
      - kubernetes
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"